Make WordIterator and Snippet::MatchPositions use size_t instead of int for offsets into strings.  This avoids some casts.  I also added a typedef for Snippet::MatchPosition which cleans up a bit of the calling code a little. Review URL: http://codereview.chromium.org/13064
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@6260 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/autocomplete/history_contents_provider.cc
@@ -218,7 +218,7 @@ void HistoryContentsProvider::ClassifyDescription(
     // Classify matches in the title.
     for (Snippet::MatchPositions::const_iterator i = title_matches.begin();
          i != title_matches.end(); ++i) {
-      if (static_cast<size_t>(i->first) != offset) {
+      if (i->first != offset) {
         match->description_class.push_back(
             ACMatchClassification(offset, ACMatchClassification::NONE));
       }
